// src/app/rma-manage/model/casadeOption.ts
export class casadeOption {
    name: string;
    code: string;
    children: object[]
}
export class children {
    cname: string;
    code: string
}


export class casadeOption2 {
    label: string;
    value: string;
    children: object[]
}
export class children2 {
    label: string;
    value: string
}